fix: saved movies after page reload
