+++ src/components/SavedMovies/SavedMovies.js
@@ -3,23 +3,51 @@ import MoviesCardList from "../MoviesCardList/MoviesCardList";
 import SearchForm from "../SearchForm/SearchForm";
 import './SavedMovies.css';
 import Preloader from '../Preloader/Preloader';
+import { CurrentUserContext } from '../../contexts/CurrentUserContext.js';
+import * as mainApi from '../../utils/MainApi';
 
 export default function SavedMovies({ 
-    getSavedMovies,
-    savedMovies,
     isLoading,
+    savedMovies,
     isServerError,
     handleRemoveMovie,
-    filterSavedMovies,
     isEmpty,
+    setIsEmpty,
+    fetchSavedMovies,
     }) {
+
+    const currentUser = React.useContext(CurrentUserContext);
 
     // short movies filter
     const [isShortOff, setIsShortOff] = React.useState(false);
 
     React.useEffect(() => {
-        getSavedMovies();
-    }, [])
+        console.log('load on saved-movies');
+        fetchSavedMovies();
+    }, [currentUser._id]);
+    
+
+    // найденные фильмы
+    const [searchedMovies, setSearchedMovies] = React.useState([]);
+
+    // фильтруем сохраненные  фильмы
+    const filterSavedMovies = (filterParam) => {
+        console.log('search');
+        const filteredMovies = savedMovies.filter((item) => {
+            const nameRU = item.nameRU.toLowerCase();
+            const nameEN = item.nameEN.toLowerCase();
+            const filter = filterParam.toLowerCase();
+
+            return nameRU.includes(filter) || nameEN.includes(filter);              
+        });
+        setSearchedMovies(
+            filteredMovies.map((item) => ({item}))
+        );
+        console.log(searchedMovies);
+        if(filteredMovies.length === 0) {
+            setIsEmpty(true);
+        }
+    }
 
     return(
         <main>
@@ -38,7 +66,8 @@ export default function SavedMovies({
                             <h2 className="movies__error">Ничего не найдено.</h2>
                             ) : (
                             <MoviesCardList
-                                movies = { savedMovies }
+                                savedMovies = { savedMovies }
+                                searchedMovies = { searchedMovies }
                                 isShortOff = { isShortOff }
                                 handleRemoveMovie = { handleRemoveMovie }
                             />
